Move generic maze model from maze-libgdx into mazeworld.generators
