+++ main.game.maze.mazeworld/src/main/java/main/game/maze/mazeworld/generators/MazeArena.java
@@ -1,0 +1,19 @@
+package main.game.maze.mazeworld.generators;
+
+import java.util.List;
+
+/**
+ * Backend-neutral view of a playable maze: a rectangular playfield, a list
+ * of wall segments, plus start and goal positions in world units.
+ */
+public interface MazeArena {
+    List<WallSegment> walls();
+
+    float widthPx();
+    float heightPx();
+
+    float startX();
+    float startY();
+    float goalX();
+    float goalY();
+}
