+++ main.game.maze.mazeworld/src/main/java/main/game/maze/mazeworld/generators/SampleMaze.java
@@ -1,4 +1,4 @@
-package main.game.maze.libgdx.game;
+package main.game.maze.mazeworld.generators;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -6,13 +6,7 @@ import java.util.List;
 import java.util.Random;
 
 /**
- * Self-contained grid maze for the libGDX gameplay subset.
- *
- * Generated with a deterministic depth-first recursive backtracker so that
- * tests can assert exact wall counts and the rendering pipeline gets the
- * same layout every run. This is a stand-in until the Tycho-packaged
- * {@code main.game.maze.mazeworld} bundle is repackaged as a plain Maven
- * jar that the libGDX module can consume directly.
+ * Self-contained grid maze generated with deterministic DFS backtracking.
  */
 public final class SampleMaze implements MazeArena {
 
@@ -37,18 +31,17 @@ public final class SampleMaze implements MazeArena {
     public int cols() { return cols; }
     public int rows() { return rows; }
     public float cellSize() { return cellSize; }
-    public float widthPx() { return cols * cellSize; }
-    public float heightPx() { return rows * cellSize; }
+    @Override public float widthPx() { return cols * cellSize; }
+    @Override public float heightPx() { return rows * cellSize; }
 
-    public List<WallSegment> walls() { return walls; }
+    @Override public List<WallSegment> walls() { return walls; }
 
-    public float startX() { return cellSize * 0.5f; }
-    public float startY() { return cellSize * 0.5f; }
-    public float goalX() { return (cols - 0.5f) * cellSize; }
-    public float goalY() { return (rows - 0.5f) * cellSize; }
+    @Override public float startX() { return cellSize * 0.5f; }
+    @Override public float startY() { return cellSize * 0.5f; }
+    @Override public float goalX() { return (cols - 0.5f) * cellSize; }
+    @Override public float goalY() { return (rows - 0.5f) * cellSize; }
 
     private static List<WallSegment> generate(int cols, int rows, float cellSize, long seed) {
-        // Each cell tracks four walls: top, right, bottom, left.
         boolean[][] top = new boolean[cols][rows];
         boolean[][] right = new boolean[cols][rows];
         boolean[][] bottom = new boolean[cols][rows];
@@ -61,7 +54,6 @@ public final class SampleMaze implements MazeArena {
 
         boolean[][] visited = new boolean[cols][rows];
         Random rnd = new Random(seed);
-        // Iterative DFS to avoid stack-overflow on big mazes.
         int[][] stack = new int[cols * rows][2];
         int sp = 0;
         stack[sp][0] = 0; stack[sp][1] = 0; sp++;
@@ -77,7 +69,6 @@ public final class SampleMaze implements MazeArena {
                 int ny = cy + (d == 0 ? -1 : d == 2 ? 1 : 0);
                 if (nx < 0 || ny < 0 || nx >= cols || ny >= rows) continue;
                 if (visited[nx][ny]) continue;
-                // Carve wall between (cx,cy) and (nx,ny).
                 switch (d) {
                     case 0: bottom[cx][cy] = false; top[nx][ny] = false; break;
                     case 1: right[cx][cy] = false; left[nx][ny] = false; break;
@@ -96,13 +87,10 @@ public final class SampleMaze implements MazeArena {
         List<WallSegment> out = new ArrayList<>();
         for (int x = 0; x < cols; x++) {
             for (int y = 0; y < rows; y++) {
-                // libGDX origin is bottom-left; row 0 sits at the bottom.
                 float x0 = x * cellSize;
                 float y0 = y * cellSize;
                 float x1 = (x + 1) * cellSize;
                 float y1 = (y + 1) * cellSize;
-                // Only emit each shared wall once: bottom + left per cell, plus
-                // the outer top/right border on the boundary cells.
                 if (bottom[x][y]) out.add(new WallSegment(x0, y0, x1, y0));
                 if (left[x][y])   out.add(new WallSegment(x0, y0, x0, y1));
                 if (y == rows - 1 && top[x][y]) out.add(new WallSegment(x0, y1, x1, y1));
